+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## UNRELEASED
 
+BREAKING CHANGES:
+- Remove automatic scheduler job enqueueing in migrations [#549](https://github.com/hlascelles/que-scheduler/pull/549)
 - Drop support for Que 0.x and Que 1.x [#548](https://github.com/hlascelles/que-scheduler/pull/548)
+
 - Dependency updates and quality improvements [#534](https://github.com/hlascelles/que-scheduler/pull/534)
 
 ## 5.1.1  (2024-08-21)